JDBC DB코드 + int id를 long id로 변경

+++ src/main/java/demo/demo/domain/Product.java
@@ -3,7 +3,7 @@ package demo.demo.domain;
 public class Product {
     private Long Id;
     private String name;
-    private int price;
+    private Long price;
     private String category;
 
     public Long getId() {
@@ -22,11 +22,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
